no-coolness
I just need to mention things here that these commits won't mean anything to me in future.
And for that I just need to mention things very explicitly and you should favour that too.
Also, don't ever be so disregardful about the things you don't have.

I would like to thank the creator of the software, it is very thoughtful of you to do that.

~Sahil Rajput